# openapi/swagger.yaml
components:
  schemas:
...
    internal_httpapi.ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    internal_httpapi.TrashListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/sudoStream_internal_trash.Item'
          type: array
          uniqueItems: false
      type: object
    internal_httpapi.TrashSettingsResponse:
      properties:
        retentionDays:
          type: integer
      type: object
...
    sudoStream_internal_trash.Item:
      properties:
        deletedAt:
          type: string
        deletedBy:
          type: string
        id:
          type: string
        libraryId:
          type: string
        originalRelPath:
          type: string
        trashRelPath:
          type: string
      type: object
...
info:
  description: HTTP API for browsing and streaming media files.
  title: sudoStream API
  version: 0.35.0
openapi: 3.1.0
paths:
...
  /api/admin/trash:
    delete:
      responses:
        "204":
          description: No Content
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Service Unavailable
      summary: Empty recycle bin
      tags:
      - admin
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.TrashListResponse'
          description: OK
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Service Unavailable
      summary: List deleted items
      tags:
      - admin
  /api/admin/trash/{id}:
    delete:
      parameters:
      - description: Trash item id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Not Found
      summary: Delete a trash item forever
      tags:
      - admin
  /api/admin/trash/{id}/restore:
    post:
      parameters:
      - description: Trash item id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_trash.Item'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Conflict
      summary: Restore a deleted item
      tags:
      - admin
  /api/admin/trash/settings:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.TrashSettingsResponse'
          description: OK
      summary: Get recycle bin settings
      tags:
      - admin
    patch:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_httpapi.TrashSettingsResponse'
                description: Recycle bin settings
                summary: body
        description: Recycle bin settings
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.TrashSettingsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
      summary: Update recycle bin settings
      tags:
      - admin
...
servers:
- description: Local development server
  url: http://localhost:8080